+++ NEWS.md
@@ -1,3 +1,5 @@
+# chromote (development version)
+
 # chromote 0.1.2
 
 # chromote 0.1.1.9001