+++ NEWS.md
@@ -2,6 +2,10 @@
 
 * `Chromote` and `ChromoteSession` gain print methods to give you a snapshot of the most important values.
 
+* `Chromote` gains a new `is_alive()` method equivalent to the old `is_active()` method; i.e. it reports on if there is an active chrome process running in the background.
+
+* Breaking change: `Chromote$is_active()` method now reports if there is an active connection to the underlying chrome instance, rather than whether or not that instance is alive (#94).
+
 * `--disable-gpu` is no longer included in the default Chrome arguments.
 
 * `ChromoteSession` now records the `targetId`. This eliminates one round-trip to the browser when viewing or closing a session. You can now call the `$respawn()` method if a session terminates and you want to reconnect to the same target (#94).