+++ NEWS.md
@@ -4,7 +4,7 @@
 
 * Breaking change: `Chromote$is_active()` method now reports if there is an active connection to the underlying chrome instance, rather than whether or not that instance is alive (#94).
 
- * `--disable-gpu` is no longer included in the default Chrome arguments, except on windows.
+* `--disable-gpu` is no longer included in the default Chrome arguments, except on windows.
 
 * `ChromoteSession` now records the `targetId`. This eliminates one round-trip to the browser when viewing or closing a session. You can now call the `$respawn()` method if a session terminates and you want to reconnect to the same target (#94).